+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 
 * Increase test coverage
 * `WebDriver#select_from_list_elements` raise an exception if entry not found
+* Remove unused `WebDriver#scroll_list_to_element` method
 
 ## 1.8.1 (2022-07-08)