+++ CHANGELOG.md
@@ -2,12 +2,17 @@
 
 ## master (unreleased)
 
+### New Features
+
+* Add chromedriver `104.0.5112.29`
+
 ### Changes
 
 * Increase test coverage
 * `WebDriver#select_from_list_elements` raise an exception if entry not found
 * Remove unused `WebDriver#scroll_list_to_element` method
 * `WebDriver#get_element_by_display` raise more detailed exception
+* Remove Chromedriver `101.0.4951.15`
 
 ## 1.8.1 (2022-07-08)